requirements.txt
langgraph
langchain
ipython
langchain_google_genai
langchain_community
python-dotenv
typing
chromadb
langchain_chroma
google-generativeai
pypdf